+++ R/sequoia.R
@@ -446,7 +446,7 @@ make_data_actions <- function(path, overwrite = FALSE) {
 
     seq_terrain(
       dirname = dirname,
-      units = "percent",
+      unit = "percent",
       overwrite = overwrite,
       verbose = verbose)
   }